[WIP]tpm: split out core TPM declarations
